Use carp to report errors from caller's perspective
Check that artist and title were provided

Carp if LWP request fails

If request returned a 404, set $Lyrics:Fetcher:Error to 'lyrics not found' 
and don't carp, as this isn't a problem, it's an expected case.

Behave nicely by using custom user agent

Set $Lyrics::Fetcher::Error when fetch() is called, so if an error occurs,
then a successful call is made, the var goes back to 'OK' rather than still
containing the last error.



git-svn-id: file:///shared/svn/scripts/perl/modules/Lyrics-Fetcher-AZLyrics@97 c1e6fe2f-0919-0410-96f0-e3221d39872c